make `pincount` in format specifier take string
In Package_Gullwing__QFP_SOIC_SO/size_definitions/soic.yaml
the `custom_name_format` format string takes a format string which will
become the name of the footprint.

`pincount` is one of the fields that can be used in this format string.
`pincount` is a string containing a number.

When the formatter is given as `{pincount:d}`, it fails, as it expects a
number but is given a string.

This commit instead formats it as a string.