A script to create Mounting Holes
It mostly follow recomendation from Daniele Moser in issue #643
(read is doc)

Footprints are fabricated in four variant

1/ a simple hole with no PAD bottom or top
   this one is the only one to have a specific
   clearance added to protect from zone filling
   type = NPTH

2t/ a hole + a PAD , top or bottom
2b/ type = NPTH

3/  a hole + PAD bottom & top
    type = PTH

4/  same as above with a dozen of vias
    connecting the PAD

The logic wants type 1 to have an inexistant PAD of
radius 0. But in KiCad, it makes it uneasy to select !
Perhaps the reason sometime they have the size of the
hole ?

It is not clear if NPTH is platting or not the hole when
you see it in 3dview

"CI" is very happy except for the 3d path inexistant
and a "virtual" thing I don't understand.

Currently 75 fp generated for metric sizes