Rework Gullwing QFN, SON pin1 silk
Use the arrow-style pin1 marker in the gullwing generator
Related: c73c8bad669d1297321491bbbfa9215bf12fd4ad

The gullwing generator creates QFP and SO* packages.
Depending on the package-style, the arrows can be in two different locations.
If the arrow can easily fit into the courtyard, its placed inside.
If the body is bigger than the pads, the arrow is

FOO better words somehow.