Merge NativeCPad into Pad
NativeCPad is an enormous copy-paste of Pad, mashed up with
some of the non-native Chamfer classes into a whole set of things
that are duplicated for no good reason.

They are also currently not used at all, so now is a good time
to do the merge before it has horrible side-effects.

Instead, do what should have been done originally, and support
chamfering of the main Pad object.

The CornerSelection class is moved into util, as it can be used
for anything, not just nodes. ChamferedPad also can use this
type. For now, keep it simple and use the same (slightly clunky)
array of bools/ints for the constructor, though this should
really be type-safer one day. But this change is hard enough to
reason about without that as well.

For now, keep ChamferedNativePad as a clone-ish of the virtual-
node ChamferedPad class. This is because ChamferedPad _is_ used
heavily and that change would have actual effects on the
generated output footprints, which will be a bit chaotic.