Terminal Block: simplify crosshead screws
The outlined cross ended up with only a tiny hair gap between
two adjacent lines, which would merge when printing, leading
to inconsistent line thickness on the physical fab prints.

Also fix a latent bug with the wrong keepouts being used
due to transforming the elements but not the keepouts.