Integrate 3D repo CI pipelines
All the 3D generator CI targets are now available as well as the
FP gen targets.

Static analysis and test stages are combined, but the generate stages
are separate, otherwise the huge number of 3D manual jobs hides the
one FP job.

Autostart of jobs is restricted based on changed directories for the
MR, so FP-only commits don't set off 3D checks and vice-versa. When
we combine YAML defs, we can adjust this as needed.

All jobs have a manual override in case you want to run them anyway
to check something if the sensitivity lists aren't picking up
changes correctly.

All the jobs are able to be 'needs: []', then they can all start in
parallel when needed.