Drawing tools: fix keepout bug
Sometimes, keepouts would be applied repeatedly to some items, which
could delete too much in some numerical edge cases. This affected, for
example, CUI terminal blocks.

Fix them, and add a couple of tests (though not exhaustive, as that
is overkill considering we have graphical diffs). But it's a handy
place to hang future regression tests if needed.

Also add some string methods (useful when the tests fail, as it
prints the parameters) and some better comments.