V9: Adjust Zone fill formatting
There actually always is a ZoneFill, even if it's none,
and KiCad (now?) outputs the thermal_gap and thermal_bridge_width
parameters mby default, even for a none fill.

Fillet radiuses of 0 and island removal mode of default (remove)
are also omitted by KiCad.

Adjust a couple of ordering changes.