More standardisation of 3D model prefixes
Conflating the libname and the 3D model directory prefix causes
wierd effects when the output directory is based on the libname.

So apply the prefix lower down, and get it from global configs in
a few places (this is part of a major campaign of GC-ification but
it's not worth doing universally just yet.