CI: Allow all generators to output to a specfied directory
Inject an environment variable to allow all generate.sh generators
(i.e. currently all of them) to know where the outputs should go.

In conjunction with the library abstraction provided by
KicadPrettyLibrary, this allows all generation results to be sent to
a specific location.

The only things not handled are the testing output of the QFN generator,
which have an output directory override in the generate.sh

Fixes: https://gitlab.com/kicad/libraries/kicad-footprint-generator/-/issues/691