Gullwing/NoLead: Disable paste via-avoid in EPs
In some (poorly-defined) circumstances, paste output from the automatic
ExposedPad paste segmnatation logic fils when paste_via_avoid is on.

Because we can't really control when it does and doesn't work, force it
OFF for all parts until it can be fixed. This overrides some YAML
settings, but means that the footprints can be regenerated without
producing unmanufacturable paste layers.

When this function of the EP is fixed (i.e. issue 674 is resolved),
this can be removed.

Relates-To: https://gitlab.com/kicad/libraries/kicad-footprint-generator/-/issues/674