Specify rounding ratio using the global configs where possible
This removes code that directly passes in round-rect parameters for
eventual assembly into RoundRadiusHandler objects. Client code should
just pass in the RoundRadiusHandler as they want. Mostly, this will
come from the global config anyway.

This thereby removes a large hanfdul of hardcodes and replaces with
the global config values. Some generators use non-default values, and
these values are kept in the interest of not causing diffs during
a technical refactor.