Pad: use round radius handler objects to control radius
Also do not mess with the pad type in the Pad class based on the
radius. The invariant is not maintained, so client code changing
the shape or radius after the fact can violate the implicit
promise that radius=0 will be a rectangle.

Instead, this logic move to the serialiser, which matches what
KiCad does and makes the code simpler before that point.