TerminalBlock: remove silkscreen screw graphics
It's nearly always kept out so much it becomes a circle and will be
hidden when assembled anyway. The silkscreen still has the outline and
alignment guides, and the pads themselves provide the obvious alignment
hints. These circles are actually quite hard to do the keepouts for
because you have to also consider the other silkscreen to avoid larger
circles crossing each other.