Connector: Drop dict-based global config, also drop some overrides
This moves the generator wholly to a GlobalConfig object interface to
global configuration data.

It also drops a few overrides in the connector YAML spec, as these are
more likely to be be set by users as global configuration rather than built
into the library definitions. They can be added back if/when there is a
specific use for overriding them for a particular series (and it's more
likely they get added back in a more 'semantic' sense for a specific
purpose.

The pad rounding handler is left in, as this feels like something
some wierd series might plausibly need for non-aesthetic reasons.