Inductor_SMD: Reinstate old Standard-Box generator
This generator has a handful of TDK inductors and one Wuerth inductor in the old standard box format. It was moved to the 3D repo by mistake and never noticed, because it was omitted from the generate.sh script.

This commit moves it back and enables it.

See merge request kicad/libraries/kicad-footprint-generator!1565