Layouts: Introduce FootprintLayout helpers
FootprintLayoutNode is a base class for a kind of Node that
represents a substantially "complete" layout for a footprint,
including:

 - Some physical geometry (pads/lines/etc), which is up to the
   inheriting class
 - Optional extras that can be opted-into by the inheritor:
    - A rectangular body outline/bound
    - A rectangular courtyard bound
    - Default texts based on those two

Generators do not have to use these, and many will not, but this
can be used by generators that want to "opt into" standard footprint
layouts for specific purposes. For example, a four-pin crystal is
functionally the same layout as a PLCC LED, just with different
parameters for the pad and body sizes.

This allows generators to defer fiddly geometry to a common core,
while still allowing them to construct the generator-specific
attributes (metadata, additional drawings, for example) on their
own terms.

In particular, generators can handle IPC or other calculations
themselves, but hand off the computed geometry to the layout for
actually turning into pads and so on.