* Crystal_SMD: Reenable rotated parts with a name suffix
This reenables the two disabled SMD crystals with an explicit rotation suffix (and adds a place in the global config to make this uniform). This is probably the easiest way out of the bind caused by that rotation.

Then we can standardise on IPC rotation level B for crystals.

See merge request kicad/libraries/kicad-footprint-generator!1588