Node: use real BoundingBox for bounding box function
This means it's of a defined type (not a dict) and can
also be directly manipulated with functions like inflate.

Also fix the Line calculateBoundingBox function, as it was
mistakenly using the "real" position, when these functions
use the node-local position. Otherwise bounding box translations
will be double counted for lines when Translations are used.