Connector: use additional_drawings rather than overloading body_shape
This allows a richer and more consistent set of drawing types, as well
as mirroring and grid repeats, that is shared with other generators
(currently: BGA, gullwing and single-row connectors)

It also allows direct control of layers (and widths/line styles if needed)
for example, you can use the 'mechanical' virtual layer if wanted.

Includes a small case where an additional drawing polygon is incorrectly
unclosed even though the first/last points are repeated.