Pad: Sanitise thermal bridge angle typing
This value always have a valid value. Apply the same defaults as KiCad,
and additionally allow a None value to mean "the default value,
whatever the pad shape" which is likely a useful thing to ask for
in the generators. Which shortcuts an annoying thing in KiCad, when
you change the pad type and get a different TBA compared to a new pad of
that shape.

Add to a serialisation tests

See: https://gitlab.com/kicad/libraries/kicad-footprint-generator/-/merge_requests/1652