Pad: fix unconnected layer handling
The serialisation was broken for remove_unused_layers (it
was always false). This replaces the two bools in Pad with the
actual information: the Padstack's unconnected layer mode,
which KiCad just happens to serialise as one or two bools.

This will also make it eaiser to move this to a Padstack
class eventually.

Add to a serialisation test.

See: https://gitlab.com/kicad/libraries/kicad-footprint-generator/-/merge_requests/1652