Pad serialisation: zone connection cannot be None
It is not meaningful for a Pad to have no zone connection value,
it must always be a value in the ZoneConnection enum.

Scrap the @property accessor, it's not adding anything here,
and just inhibits setting the connection type after construction.

Add some commentary for the inherited members and use newer Python
union syntax.

See: https://gitlab.com/kicad/libraries/kicad-footprint-generator/-/merge_requests/1652