Connector generator: refactor pad grid logic
This simplifies and genericises the way the pad grids are specified
and laid out. In theory this will allow more pad layouts, if that's
ever needed (with some YAML format surgery).

But in particular, it allow to specify a y-stagger within a single
row, which is a common idiom for lots of dual-row plug/socket connectors
that stagger each side, producing a 4-pad-row footprint of made of
a double-pitch rows, like Hirose FX2 and JAE TX24/25.

See merge request https://gitlab.com/kicad/libraries/kicad-footprint-generator/-/merge_requests/1719