Remove trailing R from names of Würth spacers
These parts were recently added by !1434
    https://gitlab.com/kicad/libraries/kicad-footprint-generator/-/merge_requests/1434
while definitions for 3D models were already added by
    https://github.com/easyw/kicad-3d-models-in-freecad/pull/392
    e998c26694fdb028c5cbc387532640a3bdef042a
in 2020 without the trailing R.

Würth recently added a trailing R to all the names of their spacers,
basically meaning just "reel version".

https://forum.digikey.com/t/9774015151r-suffix/8714

We have a lot of existing parts without the trailing R
and the names still seem to be valid, even if not always easy to find.

Also, when multiple part numbers exist whose difference is not important for PCB design,
we tend to leave out the "non-functional" part of the name.

In this case there does not seem to be a functional difference
in whether or not the trailing R is present.
All other parts that we have been shipping for years are missing the trailing R,
so it makes sense to keep the names consistent.

See merge request https://gitlab.com/kicad/libraries/kicad-footprint-generator/-/merge_requests/1724