Connector generator: allow "virtual" default specs with an fp_name
Because the name can be an evaluated pattern it's valid for a
defaults spec block to have one, even though the block itself
is not a a fully-concrete spec yet.

Instead, follow the DIP generator, and use "starts with 'defaults'
or 'base'" as the deciding factor in what is and isn't considered to
be "virtual" rather than a heuristic based on what keys exist inside
the structure.

This allows the fp_name to be defined once and filled in parametrically
by inheritors, without having the "base" definition itself always
generate parts (this can be valid, but it also can be invalid, depending
on how the YAML inheritance is structured).

Simplify a handful of definitions where possible.

See merge request https://gitlab.com/kicad/libraries/kicad-footprint-generator/-/merge_requests/1737