Connector/generator: default THT shape is oval
Having it default to circle is confusing, because the pad size
is an (x,y) vector, which doesn't work with the default shape.

If the pad size is x==y, this decays naturally to a circle anyway
on serialisation.

See merge request https://gitlab.com/kicad/libraries/kicad-footprint-generator/-/merge_requests/1744