Inductor_SMD: Port to using n-pad box layout
Use the new layout for the SMD inductor generator. This produces a
substantial decrease in the amount of code and also will allow the
generator to gain any new functions the underlying layout does. In this
case, it will support cross-wise (y) offsets of the pads, which a small
number of real Inductors have.

See: https://gitlab.com/kicad/libraries/kicad-footprint-generator/-/merge_requests/1769