PinSocket: simplify parameter loading
Nearly all the logic around making variants is redundant to information
already in, or easily expressed in, the YAML data.

So put the small amount of missing data (pitches, orientations, etc)
into the YAML, and remove the layers of parameter processing in favour of
a "normal" FootprintGenerator-compatible system.

This makes it tractable to just drop in new defintions to the YAML with
no other changes, and also makes it theoretically possible to
multi-thread by category in the future.

This commit doesn't touch the unconventional canvas-based drawing system
in socket_strips.py so as to maintain a zero-diff state.


See merge request https://gitlab.com/kicad/libraries/kicad-footprint-generator/-/merge_requests/1771